backend/requirements.txt
fastapi
uvicorn
sqlalchemy
pydantic
paho-mqtt
onnxruntime